altri refactoring
