+++ app/src/main/java/com/skycombat/game/model/gui/element/bullet/Bullet.kt
@@ -1,6 +1,6 @@
 package com.skycombat.game.model.gui.element.bullet
 
-import com.skycombat.game.model.ViewContext
+import com.skycombat.game.scene.ViewContext
 import com.skycombat.game.model.geometry.Entity
 import com.skycombat.game.model.gui.element.GUIElement
 import com.skycombat.game.model.gui.element.bullet.strategy.CollisionStrategy
