+++ app/src/main/java/com/skycombat/game/model/gui/element/enemy/Enemy.kt
@@ -1,7 +1,7 @@
 package com.skycombat.game.model.gui.element.enemy
 
 import android.graphics.*
-import com.skycombat.game.model.ViewContext
+import com.skycombat.game.scene.ViewContext
 import com.skycombat.game.model.geometry.Rectangle
 import com.skycombat.game.model.gui.Weapon
 import com.skycombat.game.model.gui.component.EnemyHealthBar
