+++ app/src/main/java/com/skycombat/game/model/gui/element/enemy/movement/Movement.kt
@@ -1,6 +1,6 @@
 package com.skycombat.game.model.gui.element.enemy.movement
 
-import com.skycombat.game.model.ViewContext
+import com.skycombat.game.scene.ViewContext
 import com.skycombat.game.model.gui.element.enemy.Enemy
 
 class Movement (private val deltaX: Int, private val deltaY: Int, private var updateFor: Int) {
