+++ app/src/main/java/com/skycombat/game/model/gui/element/powerup/PowerUp.kt
@@ -3,7 +3,7 @@ package com.skycombat.game.model.gui.element.powerup
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.PointF
-import com.skycombat.game.model.ViewContext
+import com.skycombat.game.scene.ViewContext
 import com.skycombat.game.model.geometry.Circle
 import com.skycombat.game.model.gui.element.GUIElement
 import com.skycombat.game.model.gui.element.Player
