+++ app/src/main/java/com/skycombat/game/scene/ViewContext.kt
@@ -1,14 +1,14 @@
-package com.skycombat.game.model
+package com.skycombat.game.scene
 
 class ViewContext private constructor(var width: Float, var height: Float, private var resources: android.content.res.Resources) {
     companion object{
         private var instance: ViewContext? = null
         fun setContext(width: Float, height:Float, resources: android.content.res.Resources){
-            if(instance==null) {
+            if(instance ==null) {
                 instance = ViewContext(width, height, resources)
             }
         }
-        fun getInstance():ViewContext{
+        fun getInstance(): ViewContext {
             return instance!!
         }
     }
